requirements.txt
Django==3.0.8
django-allauth==0.40.0
django-appconf==1.0.4
django-bootstrap==0.2.4
django-filter==2.4.0
django-nose==1.4.6
djangorestframework==3.12.2
Pillow==7.2.0
Pygments==2.7.2
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
django-extensions==3.1.1
django-shell-plus==1.1.7
idna==2.10
Markdown==3.3.3
nose==1.3.7
oauthlib==3.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.4
